pyproject.toml
[project]
name = "image-upload-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.12.0"
dependencies = [
    "flask==3.*",
    "functions-framework==3.*",
    "google-api-core>=2.28.1",
    "google-cloud-core>=2.5.0",
    "google-cloud-storage>=3.6.0",
]
